% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{AR1_target_params}
\alias{AR1_target_params}
\title{AR(1) process target parameters.}
\usage{
AR1_target_params(d, rho = 0.5)
}
\description{
Input:
 - d: dimension
 - rho: autocorrelation
}
\details{
Output:
 - Sigma: covariance matrix.
 - Omega: precision matrix, in sparse format.
 - Omega_chol_u: upper Choelsky factor of Omega.
}
