% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{AssembleTVDCouplingBound}
\alias{AssembleTVDCouplingBound}
\title{Assemble TVD coupling bound from meeting times "tau" at X-chain}
\usage{
AssembleTVDCouplingBound(tau, L, R, thin)
}
\description{
tau = vector of meeting times
}
\details{
L = time lag

R = number of replicates

thin = thinning factor for the bound
}
