% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{AssembleW2CouplingBound}
\alias{AssembleW2CouplingBound}
\title{Assemble W2 coupling bound.}
\usage{
AssembleW2CouplingBound(bound_comp_list, tau, L, R, thin)
}
\description{
R = number of chains
}
\details{
tau = vector of meeting times (measured at X-chain)

L = time lag

bound_comp_list = list of R vectors, each vector has entries 
||X_{t+L} - Y_t||^2

thin = thinning factor for the bound
}
