% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assignment_squared_euclidean_jackknife_cpp}
\alias{assignment_squared_euclidean_jackknife_cpp}
\title{Solve the assignment problem using TOMS1015, Guthe and Thuerck (2021), then obtain the jackknife assignment costs using the "vanilla" Jonker-Volgenant algorithm}
\usage{
assignment_squared_euclidean_jackknife_cpp(x, y, estimate_epsilon = TRUE)
}
\description{
Input: n-by-d empirical measures x and y.
}
