% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assignment_squared_euclidean_networkflow_cpp}
\alias{assignment_squared_euclidean_networkflow_cpp}
\title{Solve the assignment problem using the network simplex from Bonneel et al. (2011)}
\usage{
assignment_squared_euclidean_networkflow_cpp(x, y, compute_assignment = FALSE)
}
\description{
Input: n-by-d empirical measures x and y.
}
