% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assignment_squared_euclidean_toms1015_cpp}
\alias{assignment_squared_euclidean_toms1015_cpp}
\title{Solve the assignment problem using TOMS1015, Guthe and Thuerck (2021)}
\usage{
assignment_squared_euclidean_toms1015_cpp(x, y, estimate_epsilon = TRUE)
}
\description{
Input: n-by-d empirical measures x and y.
}
