% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{BoostrapVarianceBasic}
\alias{BoostrapVarianceBasic}
\title{"n-out-of-n" boostrap variance estimate from R repeats}
\usage{
BoostrapVarianceBasic(x, R)
}
\description{
x = vector, data to be resampled
}
\details{
R = number of bootstrap replicates
}
