% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{ComputeTranspCost1d}
\alias{ComputeTranspCost1d}
\title{One-dimensional quadratic transportation cost via order statistics}
\usage{
ComputeTranspCost1d(x, y, order_x, order_y)
}
\description{
x, y = vector of empirical measures, must be same length
}
\details{
order_x, order_y = vectors of indices, denote increasing order for data 
x and y
}
