% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EvaluateSquaredCost}
\alias{EvaluateSquaredCost}
\title{Evaluate squared Euclidean cost matrix.
     Input is assumed to be in row-major format ("n" times "d").}
\usage{
EvaluateSquaredCost(x, y)
}
\description{
Evaluate squared Euclidean cost matrix.
     Input is assumed to be in row-major format ("n" times "d").
}
