% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{EvaluateW2sqMCMC}
\alias{EvaluateW2sqMCMC}
\title{Compute W2^2 between reference measure and list of empirical measures.
(In parallel.)}
\usage{
EvaluateW2sqMCMC(x, x_reference, R, ncores)
}
\description{
x = list of matrices, each matrix represents one empirical measure
}
\details{
x_reference = matrix representing "reference" empirical measure

R = number of samples in each empirical measure

ncores = number of cores to use
}
