% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{format_singlechain_out}
\alias{format_singlechain_out}
\title{Format output for single-chain MCMC from list of R matrices of samples, where
each matrix is ("iter" x "d"), to a list of "iter" empirical measures, each 
stored as a matrix ("R" x "d").}
\usage{
format_singlechain_out(xs_list)
}
\description{
Format output for single-chain MCMC from list of R matrices of samples, where
each matrix is ("iter" x "d"), to a list of "iter" empirical measures, each 
stored as a matrix ("R" x "d").
}
