% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{GaussianMixtureDensity}
\alias{GaussianMixtureDensity}
\title{Density for mixture of Gaussians}
\usage{
GaussianMixtureDensity(x, ps, mus, sigmas)
}
\description{
x = points to be evaluated at
}
\details{
ps = mixture weights, automatically normalized to sum to 1

mus = means of mixture components

sigmas = standard deviations of mixture components
}
