% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{GenerateSigmaAndU_scaling}
\alias{GenerateSigmaAndU_scaling}
\title{Generate target variance matrix (Sigma) and its Cholesky factor (upper 
triangular; U) for the MALA vs ULA dimensional scaling experiment}
\usage{
GenerateSigmaAndU_scaling(d)
}
\description{
Matrices are sparse, stored in Matrix::dgCMatrix format.
}
