% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{GenerateTargetCovarianceMatrix_ar1}
\alias{GenerateTargetCovarianceMatrix_ar1}
\title{Generate target variance matrix for periodic boundary AR(1) experiment.}
\usage{
GenerateTargetCovarianceMatrix_ar1(d, alpha, sigma)
}
\description{
d = dimension
}
\details{
alpha = correlation in AR(1) structure

sigma = standard deviation of noise terms
}
