% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{get_empirical_w2sq_mcmc}
\alias{get_empirical_w2sq_mcmc}
\title{Calculate empirical squared Wasserstein distance in the MCMC setting}
\usage{
get_empirical_w2sq_mcmc(mcmc_out, ncores, ref_from_mcmc = T, x_ref = NA)
}
\description{
Calculate the squared empirical Wasserstein distance between hatpi_t and
hatpi_T for all t in 0,1,...T, from a list of empirical measures
list(hatpi_0, ..., hatpi_T).
}
