% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sgld_cpp}
\alias{sgld_cpp}
\title{SGLD: Stochastic gradient Langevin dynamics}
\usage{
sgld_cpp(target_params, sampler_params, theta0, iter, thin)
}
\description{
At each iteration, batches of "batch_size" are sampled uniformly without replacement. The draws are independent across iterations, i.e. we don't enforce "epochs" where we cycle over the whole dataset.
}
