% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sgld_mala_CRN_cpp}
\alias{sgld_mala_CRN_cpp}
\title{SGLD coupled with MALA (with exact gradient), CRN coupling.}
\usage{
sgld_mala_CRN_cpp(target_params, sampler_params, x0, y0, iter, thin)
}
\description{
At each iteration of SGLD, batches of "batch_size" are sampled uniformly without replacement. The draws are independent across iterations, i.e. we don't enforce "epochs" where we cycle over the whole dataset.
}
