% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sgldcv_cpp}
\alias{sgldcv_cpp}
\title{SGLD-CV: Stochastic gradient Langevin dynamics, with control variates.}
\usage{
sgldcv_cpp(target_params, sampler_params, theta0, iter, thin)
}
\description{
At each iteration, batches of "batch_size" are sampled uniformly without replacement. The draws are independent across iterations, i.e. we don't enforce "epochs" where we cycle over the whole dataset.
}
