% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SolveAssignmentNetworkflow}
\alias{SolveAssignmentNetworkflow}
\title{Solve the assignment problem using the network simplex from Bonneel et al, 2011}
\usage{
SolveAssignmentNetworkflow(C)
}
\description{
Input: n-by-n cost matrix C.
}
