% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w2sq_1d_dual_cpp}
\alias{w2sq_1d_dual_cpp}
\title{Compute 1D optimal transport cost and dual solution from sorted (equal-size) vectors of samples}
\usage{
w2sq_1d_dual_cpp(x_sorted, y_sorted)
}
\description{
The optimal dual solution is recovered by sorting the two samples, then applying  Algorithm 3 of 
https://proceedings.mlr.press/v151/sejourne22a/sejourne22a.pdf, which simplifies considerably in our case.
(There are typos in the paper: line 4 of Algorithm 3 should be "a <= b"; the initialization should be g_1 = C(x_1, y_1).
See https://github.com/thibsej/fast_uot/blob/main/fastuot/uot1d.py for a correct implementation.)
}
