% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w2sq_convergence_gaussian_recursive_cpp}
\alias{w2sq_convergence_gaussian_recursive_cpp}
\title{Evaluate W_2^2 (w/ Euclidean norm) between first d-dimensions of marginal N(mu_t[1:d], Sigma_t[1:d]) and target N(mu[1:d], Sigma[1:d])
of a process obeying the autoregression:
             (mu_{t + 1} - mu_inf) = B (mu_t - mu_inf)
             (Sigma_{t + 1} - Sigma_inf) = B (Sigma_t - Sigma_inf) B.transpose()}
\usage{
w2sq_convergence_gaussian_recursive_cpp(
  mu_0,
  Sigma_0,
  mu_inf,
  Sigma_inf,
  B,
  d,
  iter,
  thin,
  tol = 1e-12
)
}
\description{
Any vector autoregressive process with "slope" B satisfies this recurrence relation as soon as the spectrum of B is inside the ball of radius 1.
}
\details{
Deterministic Gibbs, ULA and OBAB all fall under this category.
}
