% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein.R
\name{w2sq_empirical}
\alias{w2sq_empirical}
\title{Calculate the squared empirical Wasserstein distance, with the quadratic cost, 
by solving a linear assignment problem.}
\usage{
w2sq_empirical(x, y, estimate_epsilon = TRUE)
}
\description{
Calculate the squared empirical Wasserstein distance, with the quadratic cost, 
by solving a linear assignment problem.
}
