% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein.R
\name{w2sq_empirical_1d}
\alias{w2sq_empirical_1d}
\title{One-dimensional quadratic transportation cost via order statistics
 and potentials via an O(n) post-processing step (Algorithm 3 of https://proceedings.mlr.press/v151/sejourne22a/sejourne22a.pdf)}
\usage{
w2sq_empirical_1d(x, y)
}
\description{
One-dimensional quadratic transportation cost via order statistics
 and potentials via an O(n) post-processing step (Algorithm 3 of https://proceedings.mlr.press/v151/sejourne22a/sejourne22a.pdf)
}
