% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein.R
\name{w2sq_empirical_jack}
\alias{w2sq_empirical_jack}
\title{Calculate the squared empirical Wasserstein distance, with the quadratic cost, 
by solving a linear assignment problem. Then, compute the jackknife
(i.e. leave-one-pair-of-samples-out) assignment costs.}
\usage{
w2sq_empirical_jack(x, y, estimate_epsilon = TRUE)
}
\description{
Calculate the squared empirical Wasserstein distance, with the quadratic cost, 
by solving a linear assignment problem. Then, compute the jackknife
(i.e. leave-one-pair-of-samples-out) assignment costs.
}
