% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w2sq_gaussian_cpp}
\alias{w2sq_gaussian_cpp}
\title{Evaluate W_2^2 (w/ Euclidean norm) between two multivariate Gaussians N(a, A) and N(b, B).}
\usage{
w2sq_gaussian_cpp(a, A, b, B)
}
\description{
Evaluate W_2^2 (w/ Euclidean norm) between two multivariate Gaussians N(a, A) and N(b, B).
}
