\name{wassersteinbound-package}
\alias{wassersteinbound-package}
\alias{wassersteinbound}
\docType{package}
\title{
\packageTitle{wassersteinbound}
}
\description{
\packageDescription{wassersteinbound}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wassersteinbound}
\packageIndices{wassersteinbound}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{wassersteinbound}

Maintainer: \packageMaintainer{wassersteinbound}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
